remove kazoo modify
