--- dolphinscheduler-server/src/main/java/org/apache/dolphinscheduler/server/utils/operation/NodeOperation.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.dolphinscheduler.server.utils.operation;
-
-import java.util.List;
-
-/**
- * register node operation
- */
-public interface NodeOperation {
-
-    /**
-     * judge whether node exists
-     * @param node register node
-     * @return whether node exists
-     */
-    Boolean exists(String node);
-
-
-    /**
-     * remove node
-     * @param node register node
-     * @return remove node status
-     */
-    Boolean removeNode(String node);
-
-
-    /**
-     * list nodes by path
-     * @param path path
-     * @return list nodes
-     */
-    List<String> listNodesByPath(String path);
-}
